backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
groq==0.4.1
python-multipart==0.0.6
pydantic==2.5.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
